requirements.txt
# Telegram Bot
python-telegram-bot==20.7
python-telegram-bot[ext]==20.7

# Web Framework
fastapi==0.104.1
uvicorn==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Data Processing
pandas==2.1.3
numpy==1.25.2

# R Integration
rpy2==3.5.12

# File Processing
openpyxl==3.1.2
reportlab==4.0.7

# Environment
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging
loguru==0.7.2
